Dump of port draft to Qt6/KF6 (missing Structures tool for now)
* contains the complete dump of the “it builds, starts and does not crash
  on simple usage” changes.
* continuously rebased to master branch, the latest current Qt5/KF5-based
* Structures tool disabled from build, needs QtScript port

BEWARE: not tested, expect data loss. not polished, expect broken UI.

This commit will at the time of the real Qt6/KF6 port be properly split
into the different aspects of porting.
For now it serves to hold the door open while still on the Qt6/KF6 side.

+++ kasten/gui/liboktetawidgets/bytearrayvaluestringvalidator.cpp
@@ -99,7 +99,7 @@ QString ByteArrayValueStringValidator::toString(const QByteArray& byteArray) con
 {
     QString result;
 
-    const int byteArraySize = std::min(byteArray.size(), maxLength());
+    const int byteArraySize = std::min(static_cast<int>(byteArray.size()), maxLength());
     const int encodingWidth = m_valueCodec->encodingWidth();
     result.resize(byteArraySize * encodingWidth);
     int r = 0;
